refactor(sync): extract device identity keystore into peers/identity.rs
